rename LoginService to AuthService
