+++ groclist/src/app/login/auth-guard.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
 import { Router, CanActivate } from '@angular/router';
-import { LoginService } from './login.service';
+import { AuthService } from './auth.service';
 
 @Injectable()
 export class AuthGuardService implements CanActivate {
 
-  constructor(public auth: LoginService, public router: Router) {}
+  constructor(public auth: AuthService, public router: Router) {}
 
   public canActivate(): boolean {
     if (this.auth.isAuthenticated()) {
