+++ groclist/src/app/login/login.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormBuilder } from '@angular/forms';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { LoginComponent } from './login.component';
-import { LoginService } from './login.service';
+import { AuthService } from './auth.service';
 import { JwtHelperService, JwtModule } from '@auth0/angular-jwt';
 import { RouterTestingModule } from '@angular/router/testing';
 
@@ -18,7 +18,7 @@ describe('LoginComponent', () => {
         JwtModule.forRoot({ config: { tokenGetter: () => 'token' } }),
         RouterTestingModule,
       ],
-      providers: [FormBuilder, LoginService, JwtHelperService]
+      providers: [FormBuilder, AuthService, JwtHelperService]
     })
       .compileComponents();
   }));
